Tweak headers
Hard to decide whether to have one single H1 tag for the title
then H2 tags below, or to have a H1 tag for each section.
Settling on this for now.

In the future, ideally would like to flexibly set the depth
of the TOC, thus simplifying the stylesheet. That will be
in Redcarpet 3.0 with option nesting_level, which will be supported
in Jekyll 2.0 but not before.

Or I could always write a plugin that truncates the toc at a certain
depth. But I'd rather code the Dirichlet distribution :)